move nccl comm & reduce_scatter
