+++ paddle/fluid/platform/collective_helper.h
@@ -24,6 +24,9 @@
 #include "paddle/fluid/platform/device_context.h"
 #include "paddle/fluid/platform/enforce.h"
 #include "paddle/utils/variant.h"
+#if defined(PADDLE_WITH_NCCL) || defined(PADDLE_WITH_RCCL)
+#include "paddle/fluid/distributed/collective/nccl_comm.h"
+#endif
 #if defined(PADDLE_WITH_CNCL)
 #include "paddle/fluid/platform/device/mlu/device_context.h"
 #endif
@@ -32,120 +35,8 @@ namespace paddle {
 namespace platform {
 
 #if defined(PADDLE_WITH_NCCL) || defined(PADDLE_WITH_RCCL)
-// In order to apply hierarchical communication with NCCL, we need
-// a communication ring contains NCCL communicators associated to a global
-// ncclUniqueId. E.g. for a hierarchical case,
-//
-//    11 - 12   21 - 22
-//     |    |    |    |
-//    13 - 14 - 23 - 24
-//          |    |
-//    31 - 32 - 41 - 42
-//     |    |    |    |
-//    33 - 34   43 - 44
-//
-// we group (14,23,32,41) as the top, and (11,12,13,14), (21,22,23,24),
-// (31,32,33,34), (41,42,43,44) as bottoms respectively.
-//
-// We could also use a single communication ring for the flatten case
-//
-// The NCCLComm instance is created and reversed in the NCCLCommContext
-// singleton with a global user specified group id.
-
-class NCCLComm {
- public:
-  virtual int ring_id() const = 0;
-  virtual int nranks() const = 0;
-  virtual int rank() const = 0;
-  virtual int device_id() const = 0;
-  virtual ncclComm_t comm() const = 0;
-  virtual gpuStream_t stream() const = 0;
-  virtual gpuEvent_t compute_event() const = 0;
-  virtual gpuEvent_t comm_event() const = 0;
-  virtual phi::GPUContext* dev_context() const = 0;
-  virtual ~NCCLComm() = default;
-};
-
-// A singleton NCCL communicator context reserves communication ring ids
-class NCCLCommContext {
- public:
-  static NCCLCommContext& Instance();
-
-  NCCLComm* CreateComm(
-      ncclUniqueId* nccl_id, int nranks, int rank, int dev_id, int ring_id = 0);
-
-  void CreateAllNCCLComms(const std::vector<int>& dev_ids, int ring_id = 0);
-
-  void CreateNCCLCommMultiTrainer(const std::vector<int>& dev_ids,
-                                  ncclUniqueId* nccl_id,
-                                  int nranks,
-                                  int rank,
-                                  int ring_id);
-
-  // a latter comm with the same dev_id and the same ring_id
-  // will override the former
-  NCCLComm* AssignNCCLComm(
-      ncclComm_t comm, int nranks, int rank, int dev_id, int ring_id = 0);
-
-  // retrieve a communicator by the ring id in multiprocessing mode
-  NCCLComm* Get(int ring_id) const {
-    PADDLE_ENFORCE_GT(
-        comm_map_.count(ring_id),
-        0,
-        platform::errors::InvalidArgument(
-            "Communicator in ring id %d has not been initialized.", ring_id));
-    PADDLE_ENFORCE_EQ(comm_map_.at(ring_id).size(),
-                      1,
-                      platform::errors::InvalidArgument(
-                          "One device id should be specified to retrieve from "
-                          "multiple communicators."));
-    return comm_map_.at(ring_id).begin()->second.get();
-  }
-
-  int GetRingId(ncclComm_t comm) const {
-    for (const auto& pair : comm_map_) {
-      for (const auto& p : pair.second) {
-        if (p.second.get()->comm() == comm) {
-          return pair.first;
-        }
-      }
-    }
-    return -1;
-  }
-
-  // retrieve a communicator by the ring id and the device id
-  NCCLComm* Get(int ring_id, int dev_id) const {
-    PADDLE_ENFORCE_GT(
-        comm_map_.count(ring_id),
-        0,
-        platform::errors::InvalidArgument(
-            "Communicator of ring id %d has not been initialized.", ring_id));
-    PADDLE_ENFORCE_GT(
-        comm_map_.at(ring_id).count(dev_id),
-        0,
-        platform::errors::InvalidArgument(
-            "Communicator at device id %d has not been initialized in ring %d.",
-            dev_id,
-            ring_id));
-    return comm_map_.at(ring_id).at(dev_id).get();
-  }
-
-  // retrieve a communicator by the ring id and place
-  NCCLComm* Get(int ring_id, Place place) const {
-    return Get(ring_id, place.device);
-  }
-
- private:
-  std::once_flag once_flag_;
-  std::mutex comm_map_mutex_;
-  // ring id to dev-NCCLComm
-  std::map<int, std::map<int, std::unique_ptr<NCCLComm>>> comm_map_;
-
-  void ReleaseNCCLComms();
-
-  NCCLCommContext() = default;
-  DISABLE_COPY_AND_ASSIGN(NCCLCommContext);
-};
+using NCCLComm = distributed::NCCLComm;
+using NCCLCommContext = distributed::NCCLCommContext;
 #endif
 
 #if defined(PADDLE_WITH_ASCEND_CL)
